Create api endpoint 'get-all-parcel-delivery-order'
I created a new file index.js inside the route folder and wrote the api endpoint to get all parcel order
and also wrote test for the api endpoint. I deleted two files in the model folder beacause it was not needed
anymore. I added all the information to the parcel fle in the model folder